requirements.txt
anyascii==0.3.1
asgiref==3.5.0
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==4.0.2
django-contrib-comments==2.2.0
django-filter==21.1
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
drf-yasg==1.20.0
et-xmlfile==1.1.0
generics==3.4.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
l18n==2021.3
MarkupSafe==2.0.1
openpyxl==3.0.9
packaging==21.3
pep249==1.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyTelegramBotAPI==4.1.1
python-decouple==3.5
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.2.1
TelegramBotAPI==0.3.2
telepath==0.2
template==0.7.6
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
wagtail==2.16.2
webencodings==0.5.1
Werkzeug==2.0.2
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
